Changes in domain from McAfee to Trellix
The changes are needed from December 12, 2022, at 09:30 UTC onwards due
to changes in domain. None of the API needs changes so far. Only the
region domains needs change. No change in iam auth as well. mcafee-cloud
will stay as it is. the updated script needs to be available only after
December 12, 2022, at 09:30 UTC and not before that.